requirements.txt

appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
certifi==2017.4.17
future==0.16.0
packaging==16.8
pyparsing==2.2.0
python-telegram-bot==5.3.0
requests==2.13.0
six==1.10.0
urllib3==1.20
hacker_news